Create socket folder for socketClient and socketMiddleware. Create a file for axios to run get and post calls. Add redux code to connect and disconnect from socket.io connection. Add code placeholder code to redirect user from lobby to room whenever user clicks the Create room button.
The socketClient is a class that holds the socket object itself and creates emit and on methods. The socketMiddleware takes an instance of the socketClient as an argument and is set as a middleware of the store.
Redux store middleware adds another layer of logic to the redux flow(specifically between the action and the reducer).